requirements.txt
anthropic==0.34.2
pandas==2.2.2
openpyxl==3.1.5
duckdb==1.0.0
streamlit==1.38.0
plotly==5.23.0
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.1.1
httpx==0.27.2
jinja2==3.1.4
kaleido==0.2.1
numpy==1.26.4
scipy==1.14.1
fpdf2>=2.7.9
